+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 initial release
 ## 1.1.0
 Added initial patches for IC2
+## 1.2.0
+Added tweaks for Thermal Expansion:  
+- Fuzzy NBT for recycling recipes